kaggle_requirements.txt
pandas 
numpy 
matplotlib
scipy
https://download.pytorch.org/whl/cu100/torch-1.0.1.post2-cp36-cp36m-linux_x86_64.whl
torchvision
glibc
pydicom
six
numpy
scipy
Pillow
matplotlib
scikit-image
opencv-python
imageio
Shapely
imgaug==0.2.6
albumentations
efficientnet-pytorch